project/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
assemblyai==0.42.1
certifi==2025.8.3
Django==5.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
